Modified Code - Facilitate add command.
Added code in processAdd method in Controller to create QueryList on add command to check for clashes in time with existing tasks in Model.
Added code in processAdd method in Controller to add EventObject to Model.
Fixed crash with Model.
Added code in InputParser to set milliseconds and seconds of date to zero.